requirements.txt
# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.0

# Data Visualization
plotly>=5.17.0

# Machine Learning (optional libraries for analysis)
scipy>=1.11.0
scikit-learn>=1.3.0

# LLM Integration
google-generativeai>=0.3.0

# Configuration Management
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pydantic>=2.5.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# CORS Support
python-multipart>=0.0.6

# Logging and Monitoring
structlog>=23.2.0

# File I/O
aiofiles>=23.2.0

# Date/Time utilities
python-dateutil>=2.8.0

# Environment and system
psutil>=5.9.0
